requirements.txt
numpy<2.0
pytz>=2023.3
tqdm>=4.66.0
networkx>=2.8
xxhash>=3.4
scikit-learn>=1.3,<2.0
psycopg2-binary>=2.9.9
pandas>=1.5.0
torch_geometric==2.4.0

# Reporting / dashboard / GPT
plotly>=5.22.0
streamlit>=1.32.0
openai>=1.30.0

# Note: The following are required but platform-dependent.
# Install them manually or use the provided environment.yml (recommended).
# torch
# torch-scatter
# torch-sparse
# torch-cluster
# torch-spline-conv
